Cleaned up classes and functions, separating get and print functions for all values. Verified that readings for all three phases can be sent by SMS and that readings are reliable.

+++ Phase/Phase.h
@@ -16,7 +16,7 @@ public:
 	String getName();
 	//unsigned int getPort();
 	//unsigned int getRMS();
-	unsigned int getPeriod();
+	unsigned long getPeriod();
 	signed long getDelay();
 
 	void printMessage();
